format title-fetcher codes
